+++ Server_Tic_Tac/message.h
@@ -26,29 +26,29 @@ struct Message{
     {
         QDataStream out( &data, QIODevice::WriteOnly);
         out.setVersion(QDataStream::Qt_5_3);
-        out << static_cast<qint16>(0) << QTime::currentTime() << id;
+        out << static_cast<quint16>(0) << QTime::currentTime() << id;
     }
-    Message(int v, QTcpSocket *pS = Q_NULLPTR) :
+    Message(qint16 v, QTcpSocket *pS = Q_NULLPTR) :
         id (v), pSocket (pS)
     {
         QDataStream out( &data, QIODevice::WriteOnly);
         out.setVersion(QDataStream::Qt_5_3);
-        out << static_cast<qint16>(0) << QTime::currentTime() << id;
+        out << static_cast<quint16>(0) << QTime::currentTime() << id;
     }
-    Message(int v, QByteArray byteArr, QTcpSocket *pS = Q_NULLPTR) :
+    Message(qint16 v, QByteArray byteArr, QTcpSocket *pS = Q_NULLPTR) :
         id (v), data (byteArr), pSocket (pS)
     {
 //        QDataStream out(&data, QIODevice::WriteOnly);
-//        out << static_cast<qint16>(0) << QTime::currentTime() << id;
+//        out << static_cast<quint16>(0) << QTime::currentTime() << id;
     }
-    Message(int v, QString str, QTcpSocket* pS = Q_NULLPTR):
+    Message(qint16 v, QString str, QTcpSocket* pS = Q_NULLPTR):
         id (v), pSocket (pS)
     {
         QDataStream out(&data, QIODevice::WriteOnly);
         out.setVersion(QDataStream::Qt_5_3);
-        out << static_cast<qint16>(0) << QTime::currentTime() << id << str;
+        out << static_cast<quint16>(0) << QTime::currentTime() << id << str;
     }
-    int id;
+    qint16 id;
     QTcpSocket *pSocket;
     QByteArray data;
 };
